+++ pkg/apis/admission/v1alpha1/BUILD
@@ -9,7 +9,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "doc.go",
-        "helpers.go",
         "register.go",
         "zz_generated.conversion.go",
         "zz_generated.defaults.go",
@@ -17,12 +16,10 @@ go_library(
     deps = [
         "//pkg/apis/admission:go_default_library",
         "//vendor/k8s.io/api/admission/v1alpha1:go_default_library",
-        "//vendor/k8s.io/api/authentication/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
-        "//vendor/k8s.io/apiserver/pkg/admission:go_default_library",
     ],
 )
